补充部分UT,解决部分asan问题


Match-id-5bcef115fe2eb86af59f2f79dcaec701dcf63068

+++ test/llt/unit-tests/main.cpp
@@ -13,24 +13,43 @@
 #include "test_disk.h"
 
 using namespace ock::bio;
+using namespace ock::htracer;
+
+static bool DiskPathInvalid()
+{
+    std::string filename = "./conf/bio.conf";
+    std::string target = "/dev/sdxx:/dev/sdyy";
+
+    std::ifstream file(filename);
+    std::string line;
+
+    while (getline(file, line)) {
+        if (line.find(target) != std::string::npos) {
+            return true;
+        }
+    }
+    return false;
+}
 
 int main(int argc, char *argv[])
 {
     TestCm::Stub();
     TestRpcEngine::Stub();
     TestHtracer::Stub();
-    TestDisk::Stub();
     (void) system("rm -rf conf");
     (void) system("mkdir conf");
     (void) system("cp ../configs/* conf");
     (void) system("sed -i 's/bio.mem.size_in_gb = .*/bio.mem.size_in_gb = 1/g' ./conf/bio.conf");
     (void) system("sed -i 's/bio.cm.zk_host =.*/bio.cm.zk_host = 127.0.0.1:2181/g' ./conf/bio.conf");
-    (void) system("sed -i 's/bio.disk.path = .*/bio.disk.path = test1:test2/g' ./conf/bio.conf");
+    if (DiskPathInvalid()) {
+        TestDisk::Stub();
+        (void) system("sed -i 's/bio.disk.path = .*/bio.disk.path = test1:test2/g' ./conf/bio.conf");
+        (void) system("touch test1");
+        (void) system("touch test2");
+    }
     (void) system("sed -i 's#bio.underfs.ceph.cfg.path = /etc/ceph/ceph.conf"
                   "#bio.underfs.ceph.cfg.path = ./ceph.conf#g' ./conf/bio.conf");
     (void) system("touch ceph.conf");
-    (void) system("touch test1");
-    (void) system("touch test2");
 
     auto bioServer = BioServer::Instance();
     auto ret = bioServer->Start();
@@ -41,8 +60,12 @@ int main(int argc, char *argv[])
 
     (void) system("rm -rf conf");
     (void) system("rm -rf ceph.conf");
-    (void) system("rm -rf test1");
-    (void) system("rm -rf test2");
+    if (DiskPathInvalid()) {
+        (void) system("rm -rf test1");
+        (void) system("rm -rf test2");
+    }
+
+    ClearTraceInfo();
 
     ock::htracer::HTracerExit();
     WCacheManager::Instance()->Exit();
